requirements.txt
django
gunicorn
django-heroku
django-markdownx
django-storages
boto3